+++ lib/danger/commands/local.rb
@@ -22,7 +22,6 @@ module Danger
       end
     end
 
-    # rubocop:disable Metrics/AbcSize
     def run
       ENV["DANGER_USE_LOCAL_GIT"] = "YES"
       ENV["LOCAL_GIT_PR_ID"] = @pr_num if @pr_num
@@ -67,17 +66,21 @@ module Danger
         dm.env.scm.diff_for_folder(".", from: Danger::EnvironmentManager.danger_base_branch, to: Danger::EnvironmentManager.danger_head_branch)
 
         dm.parse(Pathname.new(@dangerfile_path))
+        check_and_run_org_dangerfile(dm)
 
-        if dm.env.request_source.organisation && !dm.env.request_source.danger_repo? && (danger_repo = dm.env.request_source.fetch_danger_repo)
-          url = dm.env.request_source.file_url(repository: danger_repo.name, path: "Dangerfile")
-          path = dm.plugin.download(url)
-          dm.parse(Pathname.new(path))
-        end
         dm.print_results
       ensure
         dm.env.clean_up
       end
     end
-    # rubocop:enable Metrics/AbcSize
+
+    # Check to see if there's a Dangerfile in the organisation, and run it if so
+    def check_and_run_org_dangerfile(dm)
+      if dm.env.request_source.organisation && !dm.env.request_source.danger_repo? && (danger_repo = dm.env.request_source.fetch_danger_repo)
+        url = dm.env.request_source.file_url(repository: danger_repo.name, path: "Dangerfile")
+        path = dm.plugin.download(url)
+        dm.parse(Pathname.new(path))
+      end
+    end
   end
 end