+++ lib/danger/request_source/request_source.rb
@@ -1,6 +1,8 @@
 module Danger
   module RequestSources
     class RequestSource
+      DANGER_REPO_NAME = "danger".freeze
+
       attr_accessor :ci_source, :environment, :scm, :host, :ignored_violations
 
       def self.inherited(child_class)
@@ -42,6 +44,26 @@ module Danger
 
       def fetch_details
         raise "Subclass and overwrite initialize"
+      end
+
+      def organisation
+        raise "Subclass and overwrite organisation"
+      end
+
+      def fetch_repository(_organisation: nil, _repository: nil)
+        raise "Subclass and overwrite fetch_repository"
+      end
+
+      def fetch_danger_repo(_organisation: nil)
+        raise "Subclass and overwrite fetch_danger_repo"
+      end
+
+      def danger_repo?(_organisation: nil, _repository: nil)
+        raise "Subclass and overwrite danger_repo?"
+      end
+
+      def file_url(_organisation: nil, _repository: nil, _branch: "master", _path: nil)
+        raise "Subclass and overwrite file_url"
       end
     end
   end