+++ lib/danger/danger_core/plugins/dangerfile_git_plugin.rb
@@ -31,7 +31,6 @@ module Danger
   # @tags core, git
 
   class DangerfileGitPlugin < Plugin
-
     # The instance name used in the Dangerfile
     # @return [String]
     #