+++ lib/danger/danger_core/plugins/dangerfile_import_plugin.rb
@@ -24,7 +24,6 @@ module Danger
   # @tags core, plugins
 
   class DangerfileImportPlugin < Plugin
-
     # The instance name used in the Dangerfile
     # @return [String]
     #